Rewriting Script
I used chatGPT for like most of the orignal program because I had no clue how to get the drivers to work for the display. So now that i've finally pieced everything together i got to go back and optimize it so it isn't constantly refreshing. I want to make it display a progress bar but i dont know how the refreshing works yet.